Make upgrade mechanism.
This is a fairly simple approach. Basically every component that can mess with the rootfs on container creation can also register an onSterilise hook, which should undo what it did on creation. We then store a reference to the necessary sterilise hooks within the surrounding directory such that when things have changed, we can use the *old* sterilise script to tidy up the rootfs, before using the *new* create script to set everything else up.

Wired it all through to the systemd module stuff and it appears to work well.