Find a bug with the fixed point calculation of config.
Basically, it was sort of the same issue as the default option issue but with the whole child config being set to {} by descendPkg. Therefore we (more correctly) mix in the pkgSet.configuration to configuration before descending into the children. This is much more sensible and works.

Also by cribbing heavily off nixos/modules/tasks/network-interfaces, start fleshing out the systemd unit for the network interfaces.